Final updates as per review comments - optimise code.
